+++ src/core/colors/schemes/index.ts
@@ -1,2 +1,3 @@
 export * from './dark'
 export * from './soft-dark'
+export * from './light'