+++ src/core/colors/schemes/light.ts
@@ -1,0 +1,7 @@
+import { createAuraPalette } from '../roles'
+import { auraLightFamily } from '../source'
+import { createTemplateVars } from '../template-vars'
+
+export const light = createTemplateVars(
+  createAuraPalette(auraLightFamily, undefined, 'light')
+)