+++ src/core/colors/source/index.ts
@@ -1,2 +1,3 @@
 export * from './aura'
 export * from './variants'
+export * from './light'